refactor: extract tool turns from AgentWorkspace
